current status of work

+++ sim/common/road_state/src/road.cpp
@@ -1,4 +1,5 @@
 #include "road.h"
+#include <cstdint>
 
-Road::Road(int dest, double sL, double le) : destId(dest), speedLimit(sL), length(le) {}
+Road::Road(uint64_t dest, double sL, double le) : destId(dest), speedLimit(sL), length(le) {}
 Road::~Road() {}
